% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamp.R
\name{clamp}
\alias{clamp}
\title{Clamp Numeric Values Between a Minimum and Maximum}
\usage{
clamp(x, lower, upper)
}
\arguments{
\item{x}{A numeric vector}

\item{lower}{A single numeric value specifying the minimum allowed value.}

\item{upper}{A single numeric value specifying the maximum allowed value}
}
\value{
A numeric vector with values clamped together between \code{lower} and
\code{upper}.
}
\description{
Restricts numeric values to a specified range by replacing any values
below the lower bound with the lower bound, and any value above the
upper bound with the upper bound.
}
\examples{
clamp(c(1, 5, 10, 15, 20), lower = 5, upper = 15)
clamp(c(-2, 0, 1.5, 3.2), lower = 0, upper = 2)


}
