% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\title{Calculate Percent Change Between Two Values}
\usage{
percent_change(old, new)
}
\arguments{
\item{old}{A numeric vector of starting values.}

\item{new}{A numeric vector of ending values.}
}
\value{
A numeric vector of percent changes.
}
\description{
Calculates the percentage change from an old value to a new value.
This is a convenient alternative to writing the formula manually every time.
}
\examples{
percent_change(10, 12.5)
percent_change(25, 22)
percent_change(c(10, 25, 5), c(12.5, 22, 5.75))

}
